+++ web/src/utils.ts
@@ -570,6 +570,7 @@ export type ProtocolInfo = {
     show_model_names?: boolean,  // Show model names on top of each block (default: false)
     mqm_categories?: { [key: string]: string[] },  // Optional custom MQM categories
     mqm_severities?: string[],  // Optional custom MQM severities
+    slider_colors?: boolean,  // Optional slider colors
 }